ingest-pdf/requirements.txt
python-dotenv
pypdf
openai
weaviate-client
tiktoken
